Some cleanup around internal ethash impl
- POW tests passing
- ``get_cache()`` tests across epochs still taking a while
